data type changed from smallint to int

+++ table_geo.sql
@@ -388,7 +388,7 @@ CREATE TABLE geo.ices_area(
 CREATE TABLE geo.simple_area_cell_assignment_raw(
   id SERIAL PRIMARY KEY,
   marine_layer_id smallint NULL,
-  area_id smallint NULL,
+  area_id int NULL,
   fao_area_id smallint NULL,
   cell_id float NULL,
   water_area float NULL,
